LayoutTests/media/media-rvfc-paused-webm-expected.txt


EVENT(loadedmetadata)
EXPECTED (0 == '0') OK
EXPECTED (0 == '0') OK
EXPECTED (320 == '320') OK
EXPECTED (180 == '180') OK
EXPECTED (2.672 == '2.672') OK
EXPECTED (2.672 == '2.672') OK
EXPECTED (320 == '320') OK
EXPECTED (180 == '180') OK
